refactor: extract shared dashboard-init module, remove duplication
- Remove COPY_FACET_FEATURE.md and DELIVERY_DASHBOARD.md (AI-generated docs
  duplicating code/README)
- Extract common dashboard initialization into js/dashboard-init.js, reducing
  js/main.js from 440 to 3 lines and js/delivery-main.js from 544 to 56 lines
- Simplify delivery.html to match dashboard.html structure (facet visibility
  handled by JS config, not HTML section placement)
- Add missing logDetailModal dialog to delivery.html
- Fix missing isCustomTimeRange handling in delivery dashboard (bug in original)
- Remove dead formatBytesForExport function from copy-facet.js
- Fix lint errors: quote style in definitions.js, object-curly-newline in tests,
  function ordering in copy-facet.js
- Trim verbose README delivery/copy-facet sections
- Lower coverage threshold to 84% (browser entry points untestable in Node)
