+++ .github/workflows/preview-deploy.yml
@@ -21,6 +21,8 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v6
+        with:
+          ref: ${{ github.event.pull_request.head.sha }}
 
       - name: Set up Cloud SDK
         uses: google-github-actions/setup-gcloud@v3
@@ -32,9 +34,15 @@ jobs:
 
       - name: Deploy preview to GCS
         run: |
-          gsutil -m rsync -r -d \
+          gsutil -m -h "Cache-Control:no-cache, no-store" rsync -r -d \
             -x 'node_modules/|\.git/|\.github/|debug/|hars/|scripts/|.*\.local\.md|\.env.*|package-lock\.json|screenshot\.png' \
             . gs://klickhaus-previews/preview/pr-${{ github.event.pull_request.number }}/
+
+      - name: Invalidate CDN cache
+        run: |
+          gcloud compute url-maps invalidate-cdn-cache klickhaus-lb \
+            --path "/preview/pr-${{ github.event.pull_request.number }}/*" \
+            --project helix-225321
 
       - name: Clean up credentials
         if: always()
