+++ delivery.html
@@ -6,25 +6,10 @@
   <title>Delivery Dashboard - CDN Analytics</title>
   <link rel="canonical" href="https://klickhaus.aemstatus.net/delivery.html">
 
-  <!-- PWA Meta Tags -->
-  <meta name="apple-mobile-web-app-capable" content="yes">
-  <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
-  <meta name="apple-mobile-web-app-title" content="Delivery Dashboard">
-  <meta name="theme-color" content="#4a90d9">
-  <meta name="mobile-web-app-capable" content="yes">
-
-  <!-- Manifest & Icons -->
-  <link rel="manifest" href="/manifest.json">
+  <!-- Favicon Icons -->
   <link rel="apple-touch-icon" href="/icons/icon-180.png">
   <link rel="icon" type="image/png" sizes="32x32" href="/icons/icon-32.png">
   <link rel="icon" type="image/png" sizes="16x16" href="/icons/icon-16.png">
-
-  <!-- Service Worker Registration -->
-  <script>
-    if ('serviceWorker' in navigator) {
-      navigator.serviceWorker.register('/sw.js');
-    }
-  </script>
 
   <link rel="stylesheet" href="dashboard.css">
 </head>
@@ -92,7 +77,6 @@
         <!-- Breakdown Tables -->
         <section class="breakdowns breakdowns-pinned" id="breakdowns-pinned"></section>
         <section class="breakdowns" id="breakdowns">
-        <!-- Core Delivery Facets (visible by default) -->
         <div class="breakdown-card" id="breakdown-status-range">
           <h3>Status Range</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
@@ -106,32 +90,9 @@
           <div class="loading"><div class="spinner"></div>Loading...</div>
         </div>
         <div class="breakdown-card" id="breakdown-forwarded-hosts">
-          <h3>Forwarded Hosts (XFH)</h3>
+          <h3>Forwarded Hosts</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
         </div>
-        <div class="breakdown-card" id="breakdown-errors">
-          <h3>X-Error</h3>
-          <div class="loading"><div class="spinner"></div>Loading...</div>
-        </div>
-        <div class="breakdown-card" id="breakdown-paths">
-          <h3>Paths</h3>
-          <div class="loading"><div class="spinner"></div>Loading...</div>
-        </div>
-        <div class="breakdown-card" id="breakdown-user-agents">
-          <h3>User Agents</h3>
-          <div class="loading"><div class="spinner"></div>Loading...</div>
-        </div>
-        <div class="breakdown-card" id="breakdown-tech-stack">
-          <h3>Tech Stack</h3>
-          <div class="loading"><div class="spinner"></div>Loading...</div>
-        </div>
-        <div class="breakdown-card" id="breakdown-byo-cdn">
-          <h3>BYO CDN Type</h3>
-          <div class="loading"><div class="spinner"></div>Loading...</div>
-        </div>
-        </section>
-        <section class="breakdowns breakdowns-hidden" id="breakdowns-hidden">
-        <!-- Additional Facets (hidden by default, easy to add) -->
         <div class="breakdown-card" id="breakdown-content-types">
           <h3>Content Types</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
@@ -140,12 +101,24 @@
           <h3>Status Codes</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
         </div>
+        <div class="breakdown-card" id="breakdown-errors">
+          <h3>Errors</h3>
+          <div class="loading"><div class="spinner"></div>Loading...</div>
+        </div>
         <div class="breakdown-card" id="breakdown-cache">
           <h3>Cache Status</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
         </div>
+        <div class="breakdown-card" id="breakdown-paths">
+          <h3>Paths</h3>
+          <div class="loading"><div class="spinner"></div>Loading...</div>
+        </div>
         <div class="breakdown-card" id="breakdown-referers">
           <h3>Referers</h3>
+          <div class="loading"><div class="spinner"></div>Loading...</div>
+        </div>
+        <div class="breakdown-card" id="breakdown-user-agents">
+          <h3>User Agents</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
         </div>
         <div class="breakdown-card" id="breakdown-ips">
@@ -154,6 +127,10 @@
         </div>
         <div class="breakdown-card" id="breakdown-request-type">
           <h3>Request Types</h3>
+          <div class="loading"><div class="spinner"></div>Loading...</div>
+        </div>
+        <div class="breakdown-card" id="breakdown-tech-stack">
+          <h3>Tech Stack</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
         </div>
         <div class="breakdown-card" id="breakdown-methods">
@@ -180,6 +157,10 @@
           <h3>Cache-Control (Req)</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
         </div>
+        <div class="breakdown-card" id="breakdown-byo-cdn">
+          <h3>BYO CDN Type</h3>
+          <div class="loading"><div class="spinner"></div>Loading...</div>
+        </div>
         <div class="breakdown-card" id="breakdown-push-invalidation">
           <h3>Push Invalidation</h3>
           <div class="loading"><div class="spinner"></div>Loading...</div>
@@ -197,6 +178,7 @@
           <div class="loading"><div class="spinner"></div>Loading...</div>
         </div>
       </section>
+      <section class="breakdowns breakdowns-hidden" id="breakdowns-hidden"></section>
       </div>
 
       <!-- Logs View -->
@@ -272,7 +254,7 @@
         <div class="key-desc">Go to facet (quick navigation)</div>
 
         <div class="key-group"><kbd>r</kbd></div>
-        <div class="key-desc">Refresh data</kbd>
+        <div class="key-desc">Refresh data</div>
 
         <div class="key-group"><kbd>f</kbd></div>
         <div class="key-desc">Focus host filter</div>
@@ -298,6 +280,17 @@
         <div class="key-group"><kbd>Esc</kbd></div>
         <div class="key-desc">Exit keyboard mode</div>
       </div>
+    </div>
+  </dialog>
+
+  <!-- Log Detail Modal -->
+  <dialog id="logDetailModal" role="dialog" aria-labelledby="logDetailTitle">
+    <div class="log-detail-header">
+      <h2 id="logDetailTitle">Log Entry Details</h2>
+      <button class="modal-close" data-action="close-log-detail">&times;</button>
+    </div>
+    <div class="log-detail-content">
+      <table class="log-detail-table" id="logDetailTable"></table>
     </div>
   </dialog>
