+++ index.html
@@ -299,8 +299,14 @@
     <ul class="home-link">
       <li>
         <a href="dashboard.html">
-          <div class="title">CDN Analytics</div>
+          <div class="title">All Logs</div>
           <div class="description">Default dashboard with no filters</div>
+        </a>
+      </li>
+      <li>
+        <a href="delivery.html">
+          <div class="title">CDN Delivery</div>
+          <div class="description">Core delivery traffic only, excludes backend and admin services</div>
         </a>
       </li>
     </ul>
@@ -498,12 +504,6 @@
             title=24 Hour Overview">
             <div class="title">24 Hour Overview</div>
             <div class="description">Full day overview with top 20 values</div>
-          </a>
-        </li>
-        <li>
-          <a href="delivery.html">
-            <div class="title">Delivery Dashboard</div>
-            <div class="description">Focused delivery view with curated facets. Excludes 15 backend/admin/docs/RUM services (config, pipeline, admin, media, static, rum, content, da.live, R2 storage, docs, translate). Core delivery traffic only.</div>
           </a>
         </li>
       </ul>
