+++ js/breakdowns/definitions.js
@@ -41,13 +41,13 @@ export const allBreakdowns = [
     id: 'breakdown-status-range', col: "concat(toString(intDiv(`response.status`, 100)), 'xx')", summaryCountIf: '`response.status` >= 500', summaryLabel: 'error rate', summaryColor: 'error',
   },
   {
-    id: 'breakdown-source', col: '`source`', summaryCountIf: "`source` = 'fastly'", summaryLabel: 'fastly',
+    id: 'breakdown-source', col: '`source`', summaryCountIf: '`source` = \'fastly\'', summaryLabel: 'fastly',
   },
   {
     id: 'breakdown-hosts', col: COLUMN_DEFS.host.facetCol, linkFn: hostLink, dimPrefixes: ['main--'], summaryCountIf: "`request.host` LIKE '%.aem.live'", summaryLabel: 'live', highCardinality: true,
   },
   {
-    id: 'breakdown-forwarded-hosts', col: "`request.headers.x_forwarded_host`", linkFn: forwardedHostLink, dimFormatFn: formatForwardedHost, summaryCountIf: "`request.headers.x_forwarded_host` != ''", summaryLabel: 'production', highCardinality: true,
+    id: 'breakdown-forwarded-hosts', col: '`request.headers.x_forwarded_host`', linkFn: forwardedHostLink, dimFormatFn: formatForwardedHost, summaryCountIf: "`request.headers.x_forwarded_host` != ''", summaryLabel: 'production', highCardinality: true,
   },
   { id: 'breakdown-content-types', col: COLUMN_DEFS.contentType.facetCol, modeToggle: 'contentTypeMode' },
   { id: 'breakdown-status', col: COLUMN_DEFS.status.facetCol, modeToggle: 'contentTypeMode' },
